Flatten during report generation
Refactored the code to no longer flatten at the top-level when finishing
the top-level `ReportingTestSet` and instead perform the flattening when
we generate the JUnit XML report. Doing this allows introspecting of the
nested `ReportingTestSet` structure and opens up the possibility of
using this testset in other report generation which does support
nesting.